size_t everywhere

+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: lea <llescure@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/09 19:13:50 by lea               #+#    #+#             */
-/*   Updated: 2020/09/22 17:16:37 by lea              ###   ########.fr       */
+/*   Updated: 2020/09/23 12:12:57 by lea              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *b, int c, unsigned int len)
+#include <stdio.h>
+
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned int i;
 	char *temp;
